+++ recipes/apache-sites.rb
@@ -1,4 +1,4 @@
-node["apache"]["sites"].each do |name, site_attrs|
+node['apache']['sites'].each do |name, site_attrs|
   definition = app_vhost name do
     site site_attrs
     server_type 'apache'
@@ -10,5 +10,14 @@ node["apache"]["sites"].each do |name, site_attrs|
   else
     site = definition
   end
-  ::Chef::Mixin::DeepMerge.hash_only_merge!(node.force_override['apache']['sites'][name], site)
+
+  begin
+    values = node.attribute.combined_override['apache']['sites'][name]
+    ::Chef::Mixin::DeepMerge.hash_only_merge!(values, site)
+    node.force_override!['apache']['sites'][name] = values
+  rescue
+    # Chef 11.10 compat
+    ::Chef::Mixin::DeepMerge.hash_only_merge!(node.force_override['apache']['sites'][name], site)
+  end
+
 end